Implement 64-bit Windows generation of both MFC and wxWidgets applications using VS2010 only.
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@3814 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/core/PWHistory.cpp
@@ -22,7 +22,7 @@ bool CreatePWHistoryList(const StringX &pwh_str,
   pwh_max = num_err = 0;
 
   StringX pwh_s = pwh_str;
-  int len = pwh_s.length();
+  size_t len = pwh_s.length();
 
   if (len < 5) {
     num_err = len != 0 ? 1 : 0;
@@ -73,9 +73,9 @@ bool CreatePWHistoryList(const StringX &pwh_str,
     if (offset >= pwh_s.length())
       break;
 
-    pwh_ent.changetttdate = static_cast<time_t>(t);
+    pwh_ent.changetttdate = static_cast< time_t>(t);
     pwh_ent.changedate =
-      PWSUtil::ConvertToDateTimeString(static_cast<time_t>(t), time_format);
+      PWSUtil::ConvertToDateTimeString(static_cast< time_t>(t), time_format);
     if (pwh_ent.changedate.empty()) {
       //                       1234567890123456789
       pwh_ent.changedate = _T("1970-01-01 00:00:00");
